person view model
